[chore] Entity내부 필드 @NotBlank 및 연관관계 지연로딩 추가 - #6
