+++ dateroad-domain/src/main/java/org/dateroad/common/BaseTimeEntity.java
@@ -3,6 +3,7 @@ package org.dateroad.common;
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,8 +13,6 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.time.LocalDateTime;
-
 @Getter
 @EntityListeners(AuditingEntityListener.class)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -21,12 +20,10 @@ import java.time.LocalDateTime;
 @SuperBuilder
 @MappedSuperclass
 public abstract class BaseTimeEntity {
-
     @Column(updatable = false)
     @CreatedDate
     private LocalDateTime createdAt;
 
     @LastModifiedDate
     private LocalDateTime lastModifiedAt;
-
 }
