+++ dateroad-domain/src/main/java/org/dateroad/date/domain/Date.java
@@ -1,15 +1,17 @@
 package org.dateroad.date.domain;
 
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-import org.dateroad.common.BaseTimeEntity;
-
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
 @Entity
 @SuperBuilder
@@ -31,5 +33,4 @@ public class Date extends DateBase {
                 .startAt(startAt)
                 .build();
     }
-
 }
