+++ dateroad-domain/src/main/java/org/dateroad/dateAccess/domain/DateAccess.java
@@ -1,7 +1,18 @@
 package org.dateroad.dateAccess.domain;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.dateroad.common.BaseTimeEntity;
 import org.dateroad.date.domain.Course;
@@ -17,14 +28,17 @@ public class DateAccess extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "date_access_id", nullable = false)
+    @NotBlank
     private Long id;
 
     @OneToOne
     @JoinColumn(name = "course_id", nullable = false)
+    @NotBlank
     private Course course;
 
     @OneToOne
     @JoinColumn(name = "user_id", nullable = false)
+    @NotBlank
     private User user;
 
     public static DateAccess of(Course course, User user) {
