--- dateroad-domain/src/main/java/org/dateroad/dateTag/domain/DateTagType.java
@@ -1,6 +1,0 @@
-package org.dateroad.dateTag.domain;
-
-public enum DateTagType {
-//    TODO,~! -> date tag 작성
-    TODO,
-}
