+++ dateroad-domain/src/main/java/org/dateroad/image/domain/Image.java
@@ -1,10 +1,20 @@
 package org.dateroad.image.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
+import org.dateroad.common.BaseTimeEntity;
 import org.dateroad.date.domain.Course;
 
 @Entity
@@ -12,18 +22,19 @@ import org.dateroad.date.domain.Course;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Builder(access = AccessLevel.PRIVATE)
 @Table(name = "images")
-public class Image {
-
+public class Image extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "image_id")
     private Long id;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "course_id", nullable = false)
+    @NotBlank
     private Course course;
 
     @Column(name = "image_url", nullable = false)
+    @NotBlank
     private String imageUrl;
 
     public static Image of(Course course, String imageUrl) {
