+++ dateroad-domain/src/main/java/org/dateroad/point/domain/TransactionType.java
@@ -5,6 +5,7 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
 public enum TransactionType {
-    POINT_GAINED("획득"),POINT_USED("차감");
+    POINT_GAINED("획득"), POINT_USED("차감");
+
     private final String description;
 }
