+++ dateroad-domain/src/main/java/org/dateroad/tag/domain/DateTagType.java
@@ -1,0 +1,6 @@
+package org.dateroad.tag.domain;
+
+public enum DateTagType {
+    //    TODO,~! -> date tag 작성
+    TODO,
+}
