+++ dateroad-domain/src/main/java/org/dateroad/tag/domain/UserTag.java
@@ -1,10 +1,22 @@
-package org.dateroad.dateTag.domain;
+package org.dateroad.tag.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
+import org.dateroad.common.BaseTimeEntity;
 import org.dateroad.user.domain.User;
 
 @Entity
@@ -12,19 +24,20 @@ import org.dateroad.user.domain.User;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Builder(access = AccessLevel.PRIVATE)
 @Table(name = "user_tags")
-public class UserTag {
-
+public class UserTag extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_tag_id")
     private Long id;
 
     @JoinColumn(name = "user_id", nullable = false)
-    @ManyToOne
+    @NotBlank
+    @ManyToOne(fetch = FetchType.LAZY)
     private User user;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "date_tag", nullable = false)
+    @NotBlank
     private DateTagType dateTagType;
 
     public static UserTag of(User user, DateTagType dateTagType) {
